zibal/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
cryptography==44.0.0
Django==5.1.5
django-environ==0.12.0
django-filter==24.3
django-nine==0.2.7
djangorestframework==3.15.2
gunicorn==23.0.0
model-bakery==1.20.1
packaging==24.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pypi==2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==1.26.20
mongoengine==0.27.0
pymongo==4.5.0
khayyam
pytz
